Add Table.js from presentation :books:
